Patch/backend

+++ client/src/containers/Links/index.tsx
@@ -3,28 +3,27 @@ import LinkCard from "../../components/LinkCard";
 import Spinner from "../../components/Spinner";
 import "./stylesLinks.scss";
 import { bindActionCreators } from 'redux';
-// @ts-ignore
 import {connect} from "react-redux";
 import {loadLinks, deleteLinks} from "./actions";
 import {initialStateLinksType} from "./reducer";
 import {rootStateType} from "../../store";
 
 interface Props {
-    links:initialStateLinksType,
-    loadLinks():Promise<{}>
-    deleteLinks():void
+    links?:initialStateLinksType,
+    loadLinks?():Promise<{}>
+    deleteLinks?():void
 }
 
 const Links:React.FC<Props> = ({links, loadLinks, deleteLinks}) => {
     const [isLoading, setLoading] = useState(true);
     useEffect(() => {
-        loadLinks().then(()=>setLoading(false));
-        return () => deleteLinks();
+        if(loadLinks) loadLinks().then(()=>setLoading(false));
+        if(deleteLinks) return () => deleteLinks();
     }, [loadLinks, deleteLinks]);
     if(isLoading) return <Spinner/>;
     return (
         <div className="links">
-            {links.map(data=>
+            {links?.map(data=>
                 <LinkCard {...data}/>)}
         </div>
     )
